endpoint for spam flags
