+++ packages/backend/src/controllers/quizanswers/index.ts
@@ -13,6 +13,7 @@ import {
 import QuizService from "services/quiz.service"
 import QuizAnswerService from "services/quizanswer.service"
 import UserQuizStateService from "services/userquizstate.service"
+import ValidationService from "services/validation.service"
 import { Inject } from "typedi"
 import { EntityManager } from "typeorm"
 import { EntityFromBody } from "typeorm-routing-controllers-extensions"
@@ -32,6 +33,9 @@ export class QuizAnswerController {
   @Inject()
   private quizService: QuizService
 
+  @Inject()
+  private validationService: ValidationService
+
   @Post("/")
   public async post(@EntityFromBody() answer: QuizAnswer): Promise<any> {
     const quiz: Quiz[] = await this.quizService.getQuizzes({
@@ -47,13 +51,13 @@ export class QuizAnswerController {
       response,
       quizAnswer,
       userQuizState,
-    } = this.quizAnswerService.validateQuizAnswer(answer, quiz[0], userQState)
+    } = this.validationService.validateQuizAnswer(answer, quiz[0], userQState)
     await this.entityManager.transaction(async manager => {
       await this.userQuizStateService.createUserQuizState(
         manager,
         userQuizState,
       )
-      await this.quizAnswerService.checkForDeprecated(manager, quizAnswer)
+      await this.validationService.checkForDeprecated(manager, quizAnswer)
       await this.quizAnswerService.createQuizAnswer(manager, quizAnswer)
     })
     return response
